Add ExcludeAuthorizationHeader setting and UI controls
Introduced a new setting `GlobalSetting_ExcludeAuthorizationHeader` in `GraphEditorApplication.cs` to control the exclusion of the Authorization header from request logs.

Updated `MainEditorLoggingHandler.cs` to check this setting when processing Authorization headers, allowing them to be skipped if the setting is enabled.

Added a new toggle switch `ToggleSwitch_ExcludeAuthorizationHeader` in `GlobalSettingGeneralPage.xaml` for user control over this setting.

Implemented logic in `GlobalSettingGeneralPage.xaml.cs` to retrieve the state of the toggle switch and handle its events for saving changes.

Updated resource files `GlobalSettingGeneralPage.resw` to include descriptive text for the new toggle switch, with corresponding changes in the Japanese localization.